renamed legacy_styled back to styled
