+++ docs/src/pages/system/basics/JSS.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { legacy_styled as styled } from '@material-ui/core/styles';
+import { styled } from '@material-ui/core/styles';
 import { compose, spacing, palette } from '@material-ui/system';
 
 const Box = styled('div')(compose(spacing, palette));
