+++ packages/material-ui/src/styles/index.d.ts
@@ -25,8 +25,7 @@ export {
 } from './withStyles';
 export { default as withTheme, WithTheme } from './withTheme';
 export { default as muiStyled, CreateMUIStyled } from './muiStyled';
-export { default as legacy_styled, ComponentCreator, StyledProps } from './legacy_styled';
-export { default as styled } from './styled';
+export { default as styled, ComponentCreator, StyledProps } from './styled';
 export {
   createGenerateClassName,
   jssPreset,
