--- packages/material-ui/src/styles/legacy_styled.js
@@ -1,14 +1,0 @@
-import { styled as styledWithoutDefault } from '@material-ui/styles';
-import defaultTheme from './defaultTheme';
-
-const styled = (Component) => {
-  const componentCreator = styledWithoutDefault(Component);
-
-  return (style, options) =>
-    componentCreator(style, {
-      defaultTheme,
-      ...options,
-    });
-};
-
-export default styled;
