+++ packages/material-ui/src/styles/styled.js
@@ -1,11 +1,14 @@
-import legacy_styled from './legacy_styled';
+import { styled as styledWithoutDefault } from '@material-ui/styles';
+import defaultTheme from './defaultTheme';
 
-// @deprecated Follow the upgrade guide on http://next.material-ui.com/guides/migration-v4/#theme
 const styled = (Component) => {
-  console.error(
-    'Material-UI: This method will no longer be available in v5. Please use the legacy_styled utility instead>.',
-  );
-  return legacy_styled(Component);
+  const componentCreator = styledWithoutDefault(Component);
+
+  return (style, options) =>
+    componentCreator(style, {
+      defaultTheme,
+      ...options,
+    });
 };
 
 export default styled;
